Split http tests into multiple files
