+++ datadog_checks_base/tests/base/utils/http/test_tls_and_certs.py
@@ -1,0 +1,209 @@
+# (C) Datadog, Inc. 2019-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+import logging
+
+import mock
+import pytest
+
+from datadog_checks.base.utils.http import RequestsWrapper
+
+pytestmark = [pytest.mark.unit]
+
+
+class TestCert:
+    def test_config_default(self):
+        instance = {}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        assert http.options['cert'] is None
+
+    def test_config_cert(self):
+        instance = {'tls_cert': 'cert'}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        assert http.options['cert'] == 'cert'
+
+    def test_config_cert_and_private_key(self):
+        instance = {'tls_cert': 'cert', 'tls_private_key': 'key'}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        assert http.options['cert'] == ('cert', 'key')
+
+
+class TestIgnoreTLSWarning:
+    def test_config_default(self):
+        instance = {}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        assert http.ignore_tls_warning is False
+
+    def test_config_flag(self):
+        instance = {'tls_ignore_warning': True}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        assert http.ignore_tls_warning is True
+
+    def test_init_config_flag(self):
+        instance = {}
+        init_config = {'tls_ignore_warning': True}
+
+        http = RequestsWrapper(instance, init_config)
+
+        assert http.ignore_tls_warning is True
+
+    def test_instance_and_init_flag(self):
+        instance = {'tls_ignore_warning': False}
+        init_config = {'tls_ignore_warning': True}
+
+        http = RequestsWrapper(instance, init_config)
+
+        assert http.ignore_tls_warning is False
+
+    def test_default_no_ignore(self, caplog):
+        instance = {}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('https://www.google.com', verify=False)
+
+        expected_message = 'An unverified HTTPS request is being made to https://www.google.com'
+        for _, level, message in caplog.record_tuples:
+            if level == logging.WARNING and message == expected_message:
+                break
+        else:
+            raise AssertionError('Expected WARNING log with message `{}`'.format(expected_message))
+
+    def test_default_no_ignore_http(self, caplog):
+        instance = {}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('http://www.google.com', verify=False)
+
+        assert sum(1 for _, level, _ in caplog.record_tuples if level == logging.WARNING) == 0
+
+    def test_ignore(self, caplog):
+        instance = {'tls_ignore_warning': True}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('https://www.google.com', verify=False)
+
+        expected_message = 'An unverified HTTPS request is being made to https://www.google.com'
+        for _, _, message in caplog.record_tuples:
+            assert message != expected_message
+
+    def test_default_no_ignore_session(self, caplog):
+        instance = {'persist_connections': True}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('https://www.google.com', verify=False)
+
+        expected_message = 'An unverified HTTPS request is being made to https://www.google.com'
+        for _, level, message in caplog.record_tuples:
+            if level == logging.WARNING and message == expected_message:
+                break
+        else:
+            raise AssertionError('Expected WARNING log with message `{}`'.format(expected_message))
+
+    def test_ignore_session(self, caplog):
+        instance = {'tls_ignore_warning': True, 'persist_connections': True}
+        init_config = {}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('https://www.google.com', verify=False)
+
+        expected_message = 'An unverified HTTPS request is being made to https://www.google.com'
+        for _, _, message in caplog.record_tuples:
+            assert message != expected_message
+
+    def test_init_ignore(self, caplog):
+        instance = {}
+        init_config = {'tls_ignore_warning': True}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('https://www.google.com', verify=False)
+
+        expected_message = 'An unverified HTTPS request is being made to https://www.google.com'
+        for _, _, message in caplog.record_tuples:
+            assert message != expected_message
+
+    def test_default_init_no_ignore(self, caplog):
+        instance = {}
+        init_config = {'tls_ignore_warning': False}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('https://www.google.com', verify=False)
+
+        expected_message = 'An unverified HTTPS request is being made to https://www.google.com'
+        for _, level, message in caplog.record_tuples:
+            if level == logging.WARNING and message == expected_message:
+                break
+        else:
+            raise AssertionError('Expected WARNING log with message `{}`'.format(expected_message))
+
+    def test_instance_ignore(self, caplog):
+        instance = {'tls_ignore_warning': True}
+        init_config = {'tls_ignore_warning': False}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('https://www.google.com', verify=False)
+
+        expected_message = 'An unverified HTTPS request is being made to https://www.google.com'
+        for _, _, message in caplog.record_tuples:
+            assert message != expected_message
+
+    def test_instance_no_ignore(self, caplog):
+        instance = {'tls_ignore_warning': False}
+        init_config = {'tls_ignore_warning': True}
+        http = RequestsWrapper(instance, init_config)
+
+        with caplog.at_level(logging.DEBUG), mock.patch('requests.get'):
+            http.get('https://www.google.com', verify=False)
+
+        expected_message = 'An unverified HTTPS request is being made to https://www.google.com'
+        for _, level, message in caplog.record_tuples:
+            if level == logging.WARNING and message == expected_message:
+                break
+        else:
+            raise AssertionError('Expected WARNING log with message `{}`'.format(expected_message))
+
+
+class TestAIAChasing:
+    @pytest.mark.skip(reason="expired certified, reactivate test when certified valid again")
+    def test_incomplete_chain(self):
+        # Protocol 1.2 is allowed by default
+        http = RequestsWrapper({}, {})
+        http.get("https://incomplete-chain.badssl.com/")
+
+    def test_cant_allow_unknown_protocol(self, caplog):
+        with caplog.at_level(logging.WARNING):
+            RequestsWrapper({'tls_protocols_allowed': ['unknown']}, {})
+            assert "Unknown protocol `unknown` configured, ignoring it." in caplog.text
+        caplog.clear()
+
+    @pytest.mark.skip(reason="expired certified, reactivate test when certified valid again")
+    def test_protocol_allowed(self):
+        http = RequestsWrapper({'tls_protocols_allowed': ['TLSv1.2']}, {})
+        http.get("https://incomplete-chain.badssl.com/")
+
+    def test_protocol_not_allowed(self, caplog):
+        http = RequestsWrapper({'tls_protocols_allowed': ['TLSv1.1']}, {})
+        with caplog.at_level(logging.ERROR), pytest.raises(Exception):
+            http.get("https://incomplete-chain.badssl.com/")
+            assert "Protocol version `TLSv1.2` not in the allowed list ['TLSv1.1']" in caplog.text
